update campaign and lelang

+++ database/migrations/2020_04_25_072718_create_post_categories_table.php
@@ -17,7 +17,7 @@ class CreatePostCategoriesTable extends Migration
             $table->id();
             $table->string('name');
             $table->longText('description')->nullable();
-            $table->integer('author');
+            $table->bigInteger('author');
             $table->string('user_created');
             $table->string('user_updated')->nullable();
             $table->timestamps();
